# springboot-exercise-web/interface/mySwagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Api接口",
    "version" : "v1",
    "title" : "服务接口文档"
  },
  "host" : "localhost:8030",
  "basePath" : "/",
  "tags" : [ {
    "name" : "房屋数据管理接口"
  }, {
    "name" : "用户数据管理接口"
  } ],
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/house/query" : {
      "get" : {
        "tags" : [ "房屋数据管理接口" ],
        "summary" : "房屋数据查询接口",
        "description" : "",
        "operationId" : "queryHouse",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ResponseVOHouse"
            }
          }
        }
      }
    },
    "/house/queryByCondition" : {
      "get" : {
        "tags" : [ "房屋数据管理接口" ],
        "summary" : "房屋数据查询接口",
        "description" : "",
        "operationId" : "queryHouseByCondition",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ResponseVOListHouse"
            }
          }
        }
      }
    },
    "/house/save" : {
      "post" : {
        "tags" : [ "房屋数据管理接口" ],
        "summary" : "房屋数据保存接口",
        "description" : "",
        "operationId" : "saveUser",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/HouseVO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ResponseVOLong"
            }
          }
        }
      }
    },
    "/user/queryUser" : {
      "get" : {
        "tags" : [ "用户数据管理接口" ],
        "summary" : "查询用户数据",
        "description" : "",
        "operationId" : "queryUser",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "id",
          "in" : "query",
          "description" : "用户id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ResponseVOUser"
            }
          }
        }
      }
    },
    "/user/queryUserByType" : {
      "get" : {
        "tags" : [ "用户数据管理接口" ],
        "summary" : "按类型查询用户数据",
        "description" : "",
        "operationId" : "queryUserByType",
        "parameters" : [ {
          "name" : "userType",
          "in" : "query",
          "required" : true,
          "type" : "string",
          "enum" : [ "APP_User", "WEB_USER", "WX_APPLET_USER" ]
        }, {
          "name" : "userType",
          "in" : "query",
          "description" : "用户类型",
          "required" : true,
          "type" : "int"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ResponseVOListUser"
            }
          }
        }
      }
    },
    "/user/save" : {
      "post" : {
        "tags" : [ "用户数据管理接口" ],
        "summary" : "用户数据保存接口",
        "description" : "",
        "operationId" : "saveUser",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UserVO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ResponseVOLong"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "ExpressInfo" : {
      "type" : "object",
      "properties" : {
        "expressType" : {
          "type" : "integer",
          "format" : "int32"
        },
        "expressCompany" : {
          "type" : "string"
        },
        "expressOrderNo" : {
          "type" : "string"
        },
        "expressRemark" : {
          "type" : "string"
        }
      }
    },
    "House" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "houseAddress" : {
          "type" : "string"
        },
        "houseInfo" : {
          "$ref" : "#/definitions/HouseInfo"
        },
        "expressInfo" : {
          "$ref" : "#/definitions/ExpressInfo"
        },
        "communityName" : {
          "type" : "string"
        },
        "streetName" : {
          "type" : "string"
        },
        "createTime" : {
          "type" : "string",
          "format" : "date-time"
        },
        "updateTime" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    },
    "HouseInfo" : {
      "type" : "object",
      "properties" : {
        "bedroom" : {
          "type" : "integer",
          "format" : "int32"
        },
        "kitchen" : {
          "type" : "integer",
          "format" : "int32"
        },
        "toilet" : {
          "type" : "integer",
          "format" : "int32"
        },
        "balcony" : {
          "type" : "integer",
          "format" : "int32"
        },
        "livingRoom" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "HouseVO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "房屋主键"
        },
        "houseAddress" : {
          "type" : "string",
          "description" : "房屋地址"
        },
        "houseInfo" : {
          "description" : "房屋格局信息",
          "$ref" : "#/definitions/HouseInfo"
        },
        "expressInfo" : {
          "description" : "快递信息",
          "$ref" : "#/definitions/ExpressInfo"
        },
        "communityName" : {
          "type" : "string",
          "description" : "小区名字"
        },
        "streetName" : {
          "type" : "string",
          "description" : "街道名字"
        },
        "createTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "创建时间"
        },
        "updateTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "更新时间"
        }
      }
    },
    "ResponseVO" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        },
        "data" : {
          "type" : "object"
        }
      }
    },
    "ResponseVOHouse" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        },
        "data" : {
          "$ref" : "#/definitions/House"
        }
      }
    },
    "ResponseVOListHouse" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        },
        "data" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/House"
          }
        }
      }
    },
    "ResponseVOListUser" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        },
        "data" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/User"
          }
        }
      }
    },
    "ResponseVOLong" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        },
        "data" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "ResponseVOUser" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        },
        "data" : {
          "$ref" : "#/definitions/User"
        }
      }
    },
    "User" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "userNick" : {
          "type" : "string"
        },
        "loginName" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "tags" : {
          "type" : "array",
          "items" : {
            "type" : "object"
          }
        },
        "avatarUrl" : {
          "type" : "string"
        },
        "createTime" : {
          "type" : "string",
          "format" : "date-time"
        },
        "updateTime" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    },
    "UserVO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "主键id"
        },
        "userNick" : {
          "type" : "string",
          "description" : "用户昵称"
        },
        "loginName" : {
          "type" : "string",
          "description" : "登录名"
        },
        "password" : {
          "type" : "string",
          "description" : "登录密码"
        },
        "description" : {
          "type" : "string",
          "description" : "用户描述"
        },
        "tags" : {
          "type" : "array",
          "description" : "用户标签",
          "items" : {
            "type" : "string"
          }
        },
        "avatarUrl" : {
          "type" : "string",
          "description" : "头像地址"
        },
        "createTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "创建时间"
        },
        "updateTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "更新时间"
        }
      }
    }
  }
}